fix matching mapped value bug
