+++ src/__mocks__/@microsoft/sp-loader.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Laserfiche.
+// Licensed under the MIT License. See LICENSE.md in the project root for license information.
+
+module.exports = {
+  SPComponentLoader: {
+    loadCss: jest.fn(),
+    loadScript: jest.fn(),
+  },
+};
